backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
python-dotenv
requests
sentence-transformers
numpy
